Started on icalparser that will generate the objects based on Event.java
