+++ app/src/main/java/com/example/kthcalender/calender/Event.java
@@ -1,9 +1,11 @@
+package com.example.kthcalender.calender;
+
 public class Event {
     //TODO
     // Complete elements of a event (Föreläsning, Övning, etc..)
     public String event_type;
-    public String localization;
-    public int date; //YYYYMMDD. Maybe split into seperate?
+    public String location;
+    public int date; //YYYYMMDD.
     public int start; //Time of day
     public int end;
 }
